Improved filter handling (for lists)
- Handles null/undef more gracefully.
- More relaibly seeks out the property name in a filter.
- FilterModals are powered by store state instead of local state
- Fix range update setting restoration
